requirements.txt
beautifulsoup4==4.11.1
black==22.10.0
browser-cookie3==0.17.0
chromedriver_autoinstaller==0.4.0
colorlog==6.7.0
extract_msg==0.38.4
mypy==0.990
packaging==21.3
pandas==1.5.1
pdfplumber==0.7.6
pymongo==4.3.2
python-dotenv==0.21.0
pywin32==305
requests==2.28.1
selenium==4.6.0
synology_api==0.4.4
tqdm==4.64.1
webdriver_manager==3.8.4